reddit_streamer/requirements.txt
praw==7.5.0
requests==2.26.0
python-dotenv==0.19.2
tiktoken==0.2.0
nltk==3.6.5